requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
google-genai>=1.10.0
httpx>=0.28.0
asyncpg>=0.29.0
redis>=5.0.0
pydantic-settings>=2.2.0
PyJWT>=2.8.0
bcrypt>=4.1.0
python-multipart>=0.0.18
